fix(stage_c): update meshBias=1 (valid integer range 1-10) for Candidate MM in P43REM4-BATCH1-FWD10

+++ scripts/validation/run_real_abaqus_f43rem4_probes.py
@@ -141,7 +141,7 @@ def main():
                 sizingMethod=MINIMUM_MAXIMUM,
                 maxSolutionErrorTarget=float(rule_cfg["maxSolutionErrorTarget"]),
                 minSolutionErrorTarget=float(rule_cfg["minSolutionErrorTarget"]),
-                meshBias=float(rule_cfg["meshBias"]),
+                meshBias=int(rule_cfg["meshBias"]),
                 specifyMinSize=ON if rule_cfg["specifyMinSize"] else OFF,
                 minElementSize=float(rule_cfg["minElementSize"]),
                 specifyMaxSize=ON if rule_cfg["specifyMaxSize"] else OFF,
